lib: move shutdown.jsx and servertime.js to a lib
