+++ pkg/lib/cockpit-components-shutdown.scss
